:art: [THINGY-1] Rename _id to id field
