+++ src/main/java/com/thingy/repository/TodoRepository.java
@@ -1,9 +1,7 @@
 package com.thingy.repository;
 
-import com.thingy.model.Todo;
-import org.bson.types.ObjectId;
+import com.thingy.entity.Todo;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface TodoRepository extends MongoRepository<Todo, String> {
-  Todo findBy_id(ObjectId id);
 }
